mapper personOpplysninger.
