+++ src/main/kotlin/no/nav/tilleggsstonader/sak/behandling/fakta/BehandlingFaktaUtils.kt
@@ -1,6 +1,6 @@
 package no.nav.tilleggsstonader.sak.behandling.fakta
 
-import no.nav.tilleggsstonader.sak.opplysninger.søknad.boutgifter.Personopplysninger
+import no.nav.tilleggsstonader.sak.opplysninger.søknad.domain.Personopplysninger
 
 fun mapPersonopplysninger(personopplysninger: Personopplysninger?): FaktaPersonopplysninger =
     FaktaPersonopplysninger(
