+++ src/main/kotlin/no/nav/tilleggsstonader/sak/opplysninger/søknad/boutgifter/SkjemaBoutgifter.kt
@@ -3,6 +3,7 @@ package no.nav.tilleggsstonader.sak.opplysninger.søknad.boutgifter
 import no.nav.tilleggsstonader.kontrakter.søknad.JaNei
 import no.nav.tilleggsstonader.sak.opplysninger.søknad.domain.AktivitetAvsnitt
 import no.nav.tilleggsstonader.sak.opplysninger.søknad.domain.HovedytelseAvsnitt
+import no.nav.tilleggsstonader.sak.opplysninger.søknad.domain.Personopplysninger
 import java.time.LocalDate
 
 data class SkjemaBoutgifter(
@@ -12,18 +13,6 @@ data class SkjemaBoutgifter(
     val boutgifter: BoligEllerOvernattingAvsnitt,
     val dokumentasjon: List<DokumentasjonBoutgifter>,
     val harNedsattArbeidsevne: JaNei?,
-)
-
-data class Personopplysninger(
-    val adresse: Adresse?,
-)
-
-data class Adresse(
-    val gyldigFraOgMed: LocalDate?,
-    val adresse: String?,
-    val postnummer: String?,
-    val poststed: String?,
-    val landkode: String?,
 )
 
 data class BoligEllerOvernattingAvsnitt(
